moved getAllEvents code into function, wrote getEventsSince func, still need to test
